requirements.txt
albumentations==0.5.2
comet_ml==3.2.3
pretrainedmodels==0.7.4
efficientnet_pytorch==0.6.3
torch==1.7.0
omegaconf==2.0.5
pytorch_toolbelt==0.4.0
hydra_core==1.0.4
hydra_colorlog==1.0.0
pytest==6.1.2
numpy==1.19.4
pandas==1.1.4
torchvision==0.8.1
pytorch_lightning==1.1.0
Pillow==8.0.1
scikit_learn==0.23.2
adabound==0.0.5
adabelief-pytorch==0.1.0
ranger-adabelief==0.1.0
timm==0.3.1
